+++ handlers/novo/novo.go
@@ -6,12 +6,13 @@ import (
 	"net/http"
 	"strings"
 
+	"net/url"
+	"time"
+
+	"github.com/buger/jsonparser"
 	"github.com/nyaruka/courier"
 	"github.com/nyaruka/courier/handlers"
 	"github.com/nyaruka/courier/utils"
-	"net/url"
-	"github.com/buger/jsonparser"
-	"time"
 )
 
 const (
@@ -78,8 +79,8 @@ func (h *handler) receiveMessage(ctx context.Context, c courier.Channel, w http.
 
 // SendMsg sends the passed in message, returning any error
 func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStatus, error) {
-	merchantId := msg.Channel().StringConfigForKey(configMerchantId, "")
-	if merchantId == "" {
+	merchantID := msg.Channel().StringConfigForKey(configMerchantId, "")
+	if merchantID == "" {
 		return nil, fmt.Errorf("no merchant_id set for NV channel")
 	}
 
@@ -89,10 +90,10 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	}
 
 	status := h.Backend().NewMsgStatusForID(msg.Channel(), msg.ID(), courier.MsgErrored)
-	parts := handlers.SplitMsg(handlers.GetTextAndAttachments(msg), maxMsgLength)
+	parts := handlers.SplitMsgByChannel(msg.Channel(), handlers.GetTextAndAttachments(msg), maxMsgLength)
 	for _, part := range parts {
 		from := strings.TrimPrefix(msg.Channel().Address(), "+")
-		to   := strings.TrimPrefix(msg.URN().Path(), "+")
+		to := strings.TrimPrefix(msg.URN().Path(), "+")
 
 		form := url.Values{
 			"from": []string{from},
@@ -101,7 +102,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 		}
 		form["signature"] = []string{utils.SignHMAC256(merchantSecret, fmt.Sprintf("%s;%s;%s;", from, to, part))}
 
-		partSendURL, _ := url.Parse(fmt.Sprintf(sendURL, merchantId))
+		partSendURL, _ := url.Parse(fmt.Sprintf(sendURL, merchantID))
 		partSendURL.RawQuery = form.Encode()
 
 		req, _ := http.NewRequest(http.MethodGet, partSendURL.String(), nil)