+++ handlers/thinq/thinq.go
@@ -183,7 +183,7 @@ func (h *handler) SendMsg(_ context.Context, msg courier.Msg) (courier.MsgStatus
 
 	// now send our text if we have any
 	if msg.Text() != "" {
-		parts := handlers.SplitMsg(msg.Text(), maxMsgLength)
+		parts := handlers.SplitMsgByChannel(msg.Channel(), msg.Text(), maxMsgLength)
 		for _, part := range parts {
 			body := mtMessage{
 				FromDID: strings.TrimLeft(msg.Channel().Address(), "+")[1:],